LPS-91241 Pull bundle creation logic to BndBundleUtil
